Update scraping code after hotel has changed booking system
Old way still works but gives prices higher than shown when manually looking on the website. This change coincides with an apparent price drop through the board but that seems to check out. Updates scraper still relies on the old method for determining how many instances of a room are available, since the new booking system doesn't expose this directly.

This commit also makes a switch go gather data in English, where applicable.